Itemlist (wo restrictions) as comma separated list
Previously itemlists wo restrictions were rendered as repeats. Problem
is that repeats do not allow to enter default values. Also a select is
no option since the user can't add values

So I suggest to use a text field allowing for comma separated list of
strings, integers, or floats. The correctness of the input is verified
by by validators and sanitizers.

I'm optimistic that listitems do not contain ',' (which would be a
problem), since restrictions of restricted itemslists are given as comma
separated list.