added code to autogenerate a simple test for unit tests
idea: take the generated input and output nodes and adapt it to a test
case. mainly the optional parameters and outputs are stripped and
attributes changes appropriately.

adds planemo linting and testing to the unit tests

for the future this shall be extended to take a ctd generated by the
openms unit tests. the problem with the unit tests is that only files
are given on the command line, but all other parameters in a ctd file.
using an improved version of this tool the following strategy seems
useful: generate the full xml file and extract just the test part.

also fixes a few minor errors detected with the test