some fixes for galaxy (#57)
* some fixes for galaxy selects

radio selects are used for non-list (`multiple='false'` in the Galaxy world)
choices with less than 5 possible options.

those are incompatible with `required=true` (`optional="false"` in the Galaxy world).

when respecting this also the artificial `default (nothing chosen)`
option is not needed anymore.

* change to checkboxes

which work for all kinds of selects

* improve optional attribute of parameters

OpenMS uses sets text, int, select, bool parameters that have a default
as optional (required=False), the default value is set implicitly if no
value is given.
This is reasonable for the CLI because one certainly does not want the
user to specify the default manually for all parameters.
For Galaxy tools setting these parameters as required leads to the
equivalent behavior. Assuming required is better because it makes
the implicit setting of parameters more transparent to the user
(in Galaxy the default would be prefilled in the form and at least
one option needs to be selected).